backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.7
jinja2==3.1.2
httpx==0.24.1
EOL